:watch: Add opening times
Pull opening times for GP practice from syndication and display on the
GP page.
Much refactoring to do to remove duplication in code and tests.
Need to think more about the structure of the opening times data
as at the moment it is tightly coupled to the syndication data
structure.