:wine: Handle missing opening times
Code previously would ignore pharmacies without opening times.

Needed to change the tests as mocking was set up for 10 responses to
the pharamcy overview url. After 10 requests null was returned which
was causing the tests to fail with a null reference exception which was
masked by the previous implementation.

The tests are now quite slow (even with mocking) as there are 100
requests to the pharmacy oveview url.