:articulated_lorry: move backstop tests
As it stood mocha was trying to run the backstop tests. Explored using
an exclude switch for mocha but having separate subdirectories for mocha
and backstop seems a cleaner solution since it doesn't involve making
changes to the many mocha and istanbul yarn commands.  Unfortunately
there is a lot of work going on in the feature/azure-search-wrap branch
which would result in many merge conflicts if we move the mocha tests
into a subdirectory now.  Moving the backstop tests up to the parent is
a temporary step until the feature branch work has been completed.

The proposed future directory structure is:
- test
  - mocha
  - backstop