:hammer: :bug: Correct for error in data
Additional opening times for phamacies (and possibly other org types)
can have a type of _either_ Additional or General.  This is an error in
the source data which seems to have been introduced in 2013 and never
corrected.  We have extended the guard rather than remove it (since
'additionalness' can be infered by a non-zero AdditionalOpeningDate) to
filter out any future opening times for different types (e.g. reception)