:bar_chart: change analytics base data layer
Ahead of the move to live from beta the data layers used in analytics
need to change. It is OK to make this change now while analytics are
not being collected on the assumption that this will be live before the
end of Oct. If not then we will have to switch it back to it's previous
value.